+++ Bencodex/Decoder.cs
@@ -28,27 +28,24 @@ namespace Bencodex
 
         public IValue Decode()
         {
-            IValue value = DecodeValue();
-            if (ReadByte() is { } b)
-            {
+            IValue value = DecodeValue() ??
                 throw new DecodingException(
-                    $"An unexpected trailing byte 0x{b:x} at {_offset - 1}."
-                );
-            }
+                    $"An unexpected token byte 0x{0x65:x} at {_offset - 1}");
 
-            return value;
+            return EndOfStream()
+                ? value
+                : throw new DecodingException(
+                    $"An unexpected trailing byte remains at {_offset}.");
         }
 
-        private IValue DecodeValue()
+        private IValue? DecodeValue()
         {
             const byte e = 0x65;  // 'e'
 
             switch (ReadByte())
             {
-                case null:
-                    throw new DecodingException(
-                        $"The byte stream terminates unexpectedly at {_offset}."
-                    );
+                case 0x65: // 'e'
+                    return null;
 
                 case 0x6e: // 'n'
 #pragma warning disable SA1129
@@ -70,18 +67,8 @@ namespace Bencodex
 
                 case 0x6c: // 'l'
                     var elements = new List<IValue>();
-                    while (true)
+                    while (DecodeValue() is IValue element)
                     {
-                        byte b = ReadByte() ?? throw new DecodingException(
-                            $"The byte stream terminates unexpectedly at {_offset}."
-                        );
-                        if (b == e)
-                        {
-                            break;
-                        }
-
-                        Back();
-                        IValue element = DecodeValue();
                         elements.Add(element);
                     }
 
@@ -89,19 +76,11 @@ namespace Bencodex
 
                 case 0x64: // 'd'
                     var pairs = new List<KeyValuePair<IKey, IValue>>();
-                    while (true)
+                    while (DecodeKey() is IKey key)
                     {
-                        byte b = ReadByte() ?? throw new DecodingException(
-                            $"The byte stream terminates unexpectedly at {_offset}."
-                        );
-                        if (b == e)
-                        {
-                            break;
-                        }
-
-                        Back();
-                        IKey key = DecodeKey();
-                        IValue value = DecodeValue();
+                        IValue value = DecodeValue()
+                            ?? throw new DecodingException(
+                                $"An unexpected token byte 0x{0x65:x} at {_offset - 1}");
                         pairs.Add(new KeyValuePair<IKey, IValue>(key, value));
                     }
 
@@ -125,14 +104,12 @@ namespace Bencodex
             }
         }
 
-        private IKey DecodeKey()
+        private IKey? DecodeKey()
         {
             switch (ReadByte())
             {
-                case null:
-                    throw new DecodingException(
-                        $"Expected a dictionary key, but the byte stream terminates at {_offset}."
-                    );
+                case 0x65: // 'e'
+                    return null;
 
                 case 0x75: // 'u':
                     return ReadTextAfterPrefix();
@@ -184,7 +161,7 @@ namespace Bencodex
             return buffer;
         }
 
-        private byte? ReadByte()
+        private byte ReadByte()
         {
             if (_didBack)
             {
@@ -201,7 +178,20 @@ namespace Bencodex
 
             _offset++;
             _didBack = false;
-            return read == 0 ? (byte?)null : _tinyBuffer[0];
+            return read == 0
+                ? throw new DecodingException($"The byte stream terminates unexpectedly at {_offset}.")
+                : _tinyBuffer[0];
+        }
+
+        // Checks end of stream.  Should be called only once at the very end.
+        private bool EndOfStream()
+        {
+            if (_didBack)
+            {
+                return false;
+            }
+
+            return _stream.Read(_tinyBuffer, 0, 1) == 0;
         }
 
         private void Back()
@@ -224,15 +214,7 @@ namespace Bencodex
             const int asciiZero = 0x30; // '0'
             int length = 0;
 
-            var b = ReadByte();
-
-            if (b is null)
-            {
-                throw new DecodingException(
-                    $"Expected digits, but the byte stream terminates at {_offset}.");
-            }
-
-            byte lastByte = b.Value;
+            byte lastByte = ReadByte();
             while (lastByte != colon)
             {
 #pragma warning disable SA1131
@@ -247,10 +229,7 @@ namespace Bencodex
                 length *= 10;
                 length += lastByte - asciiZero;
 
-                lastByte = ReadByte() ?? throw new DecodingException(
-                    $"Expected a delimiter byte 0x{colon:x}, but the byte stream terminates " +
-                    $"at {_offset}."
-                );
+                lastByte = ReadByte();
             }
 
             return length;
@@ -262,27 +241,11 @@ namespace Bencodex
             byte[] buffer = new byte[defaultBufferSize];
 
             var b = ReadByte();
-
-            if (b is null)
-            {
-                throw new DecodingException(
-                    $"Expected a minus sign or a digit, " +
-                    $"but the byte stream terminates at {_offset}."
-                );
-            }
-
             bool minus = false;
             if (b == 0x2d) // '-'
             {
                 minus = true;
                 b = ReadByte();
-
-                if (b is null)
-                {
-                    throw new DecodingException(
-                        $"Expected digits, but the byte stream terminates at {_offset}."
-                    );
-                }
             }
 
             int digitsLength;
@@ -290,16 +253,16 @@ namespace Bencodex
             if (minus)
             {
                 buffer[0] = 0x2d;
-                buffer[1] = b.Value;
+                buffer[1] = b;
                 digitsLength = 2;
             }
             else
             {
-                buffer[0] = b.Value;
+                buffer[0] = b;
                 digitsLength = 1;
             }
 
-            byte lastByte = b.Value;
+            byte lastByte = b;
 
             while (lastByte != delimiter)
             {
@@ -312,10 +275,7 @@ namespace Bencodex
                 }
 #pragma warning restore SA1131
 
-                lastByte = ReadByte() ?? throw new DecodingException(
-                    $"Expected a delimiter byte 0x{delimiter:x}, but the byte stream terminates " +
-                    $"at {_offset}."
-                );
+                lastByte = ReadByte();
 
                 if (digitsLength >= buffer.Length)
                 {